Find Tumbling George
This is an assembly program from the Programming Hardware-Software
Systems class at Georgia Tech, Taught by Bo Hong. This code requires the
Misasim software to be installed in order to interpret the code. This
program is written in the MIPS flavor of assembly. This code is more
efficient than the code base written by the professor to be used as a
baseline.